chore(haproxy): Disable Feature-Policy header
Feature-Policy was causing My Oswee to fail. Probably because of
full-screen, sounds and other. To simplify, i just disabled this rule.

Yayy! Finally some working and reproducable setup. :)
I think i will continue with expanding Ansible to automate Podman
and My Oswee setup as well.
But next task is to improve the performance.